resume-parser-using-OCR/requirements.txt
numpy
scikit-learn
opencv-python
pillow
matplotlib
pdf2image
roboflow
ultralytics
pytesseract
easyocr
pdfplumber
